♻️REFACTOR: API versioning for new structure
- created a new folder to contain the V1 version of the API.

- changed itinerary routes, controllers, models and services.

- PURPOSE: for better scalability when making new changes that greatly affect the API.
