+++ src/controllers/V1/API/user/login.js
@@ -1,5 +1,5 @@
-const User = require("../../../models/user.models");
-const { generateToken, verifyToken } = require("../../../services/Tokens");
+const User = require("../../../../models/user.models");
+const { generateToken, verifyToken } = require("../../../../services/Tokens");
 const bcryptjs = require("bcryptjs");
 const login = async (req, res) => {
   const { email, password } = req.body;
